[FIXED] Default exception handler test location
